pair_programming_3/bsd-ftp-sync/requirements.txt
google-api-core==1.14.3
google-api-python-client==1.7.11
google-cloud-pubsub==1.0.2
google-auth==1.6.3
google-cloud-core==1.0.3
google-cloud-storage==1.20.0
google-cloud-logging==1.14.0
pylint==2.4.2
python-dateutil==2.8.0
python-dotenv==0.14.0
PyYAML==5.3
pysftp==0.2.9
parse==1.15.0